:hammer: [app] move component folder
