+++ src/app/src/components/ConvItem.tsx
