+++ src/app/src/components/TComponent/index.ts
