+++ src/app/src/components/messageTypes/Base.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { View, Text } from 'react-native';
 import sb from 'react-native-style-block';
 import { TAvatar } from '../TComponent';
-import dateHelper from '../../../shared/utils/date-helper';
-import config from '../../../shared/project.config';
+import dateHelper from '../../../../shared/utils/date-helper';
+import config from '../../../../shared/project.config';
 import { MessageProps } from '@shared/components/MessageHandler';
 import _get from 'lodash/get';
 import { getAbsolutePath } from '@shared/utils/file-helper';
