+++ src/app/src/components/messageTypes/Card.tsx
