+++ src/app/src/components/messageTypes/Tip.tsx
