+++ src/app/src/components/messageTypes/card/BaseCard.tsx
