+++ src/app/src/components/messageTypes/card/FriendInvite.tsx
