+++ src/app/src/screens/ChatScreen.tsx
@@ -12,19 +12,19 @@ import { Icon } from '@ant-design/react-native';
 import sb from 'react-native-style-block';
 import ImagePicker from 'react-native-image-picker';
 import { NavigationScreenProps } from 'react-navigation';
-import { TInput, TIcon } from '../../components/TComponent';
+import { TInput, TIcon } from '../components/TComponent';
 import config from '../../../shared/project.config';
 import { sendMsg } from '../../../shared/redux/actions/chat';
 import { getUserInfoCache } from '../../../shared/utils/cache-helper';
 import dateHelper from '../../../shared/utils/date-helper';
-import ExtraPanelItem from '../../components/chat/ExtraPanelItem';
-import EmotionPanel from '../../components/chat/EmotionPanel';
+import ExtraPanelItem from '../components/chat/ExtraPanelItem';
+import EmotionPanel from '../components/chat/EmotionPanel';
 import { toNetwork } from '../../../shared/utils/image-uploader';
 import { toTemporary } from '../../../shared/utils/upload-helper';
 import { unemojify } from '../utils/emoji';
 import _get from 'lodash/get';
 
-import MessageHandler from '../../components/messageTypes/__all__';
+import MessageHandler from '../components/messageTypes/__all__';
 
 import styled from 'styled-components/native';
