+++ src/app/src/screens/RegisterScreen.tsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { View, Text } from 'react-native';
 import { NavigationActions } from 'react-navigation';
 import sb from 'react-native-style-block';
-import { TButton, TFormGroup } from '../../components/TComponent';
+import { TButton, TFormGroup } from '../components/TComponent';
 import { showLoading, showAlert, hideAlert } from '../../../shared/redux/actions/ui';
 import { register } from '../../../shared/redux/actions/user';
